move shared sync_stock_transactions to somewhere actually shared
