+++ custom/ilsgateway/views.py
@@ -27,8 +27,7 @@ from custom.ilsgateway.tanzania.reminders.randr import send_ror_reminder
 from custom.ilsgateway.tanzania.reminders.stockonhand import send_soh_reminder
 from custom.ilsgateway.tanzania.reminders.supervision import send_supervision_reminder
 
-from custom.ilsgateway.tasks import sync_product_stocks, sync_stock_transactions, get_supply_point_statuses, \
-    get_delivery_group_reports, ILS_FACILITIES, get_ilsgateway_data_migrations
+from custom.ilsgateway.tasks import ILS_FACILITIES, get_ilsgateway_data_migrations
 from casexml.apps.stock.models import StockTransaction
 from custom.logistics.tasks import sms_users_fix
 from custom.ilsgateway.api import ILSGatewayAPI
